+++ src/uttaksplan/components/skjema/spørsmål/DekningsgradSpørsmål.tsx
