+++ src/uttaksplan/components/skjema/spørsmål/FordelingFellesperiodeSpørsmål.tsx
