+++ src/uttaksplan/components/skjema/spørsmål/fordelingFellesperiode/FordelingFellesperiodeLabelRenderer.tsx
