+++ src/uttaksplan/components/skjema/uttaksperiodeSkjema/uttaksperiodeskjemaUtils.ts
